requirements.txt
fastapi
uvicorn
pytest-mock